bot/requirements.txt
backports.ssl-match-hostname==3.5.0.1
bitso-py==3.0.0
certifi==2018.8.13
chardet==3.0.4
funcsigs==1.0.2
idna==2.7
mock==2.0.0
pbr==4.2.0
python-dateutil==2.7.3
requests==2.19.1
six==1.11.0
urllib3==1.23
websocket-client==0.40.0
pycron==0.80.0